Add support for anyOf and allOf features in RequiresFeature Annotations
The existing RequiresFeature system is limited to single feature
dependencies. This patch introduces anyOf and allOf feature dependency
support to RequiresFeature. This allows some of the Telephony API to be
enabled based on either a subset or the complete set of required
features.

Ignore-AOSP-First: in topic with internal-first/internal-only projects
Bug: 400781089
Flag: EXEMPT bug fix
Test: manual inspection of docs
Change-Id: I82fb169734166e1757ede6ad551d4d1f28b8a619